Revert PR #688: Return to monolithic bench_cfg.py
The modular approach added unnecessary complexity through delegation magic.
Keeping monolithic structure provides:
- Simple CLI discovery via bench_cfg.<tab>
- Full backward compatibility
- Cleaner param library integration
- Simpler mental model

The 680-line file is manageable with proper section comments.
